refactor(multi-channel): split runtime incident, inspect, and tests modules (#1230)
